+++ R/create.r
@@ -1,7 +1,7 @@
 #' Creates a new package, following all devtools package conventions.
 #'
 #' Similar to \code{\link{package.skeleton}}, except that it only creates
-#' the standard devtools directory structures, it doesn't try and create
+#' the standard devtools directory structures; it doesn't try and create
 #' source code and data files by inspecting the global environment.
 #'
 #' @param path location to create new package.  The last component of the path
@@ -28,6 +28,15 @@
 create <- function(path, description = getOption("devtools.desc"),
                          check = FALSE, rstudio = TRUE) {
   name <- basename(path)
+  if (!valid_name(name)) {
+    stop(
+      name, " is not a valid package name: it should contain only\n",
+      "ASCII letters, numbers and dot, have at least two characters\n",
+      "and start with a letter and not end in a dot.",
+      call. = FALSE
+    )
+  }
+
   message("Creating package ", name, " in ", dirname(path))
 
   if (file.exists(path)) {
@@ -46,6 +55,10 @@ create <- function(path, description = getOption("devtools.desc"),
 
   if (check) check(path)
   invisible(TRUE)
+}
+
+valid_name <- function(x) {
+  grepl("^[[:alpha:]][[:alnum:].]+$", x) && !grepl("\\.$", x)
 }
 
 create_namespace <- function(path) {