+++ inst/templates/readme-rmd-pre-commit.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+if [[ README.Rmd -nt README.md ]]; then
+  echo "README.md is out of date; please re-knit README.Rmd"
+  exit 1
+fi